requirements.txt
streamlit>=1.30
python-dotenv>=1.0
google-genai>=1.0.0
reportlab
